requirements.txtファイルの更新完了。Webサーバーが起動したら、dbが作成される設定。importエラーが出ている。一時的なコミット
