+++ .env.example
@@ -1,0 +1,4 @@
+# shellcheck shell=ksh
+# 上記コメントでShellCheckはスクリプトを正しく解釈し、適切なチェックを実行する
+# Example database URL
+DATABASE_URL=postgresql+psycopg2://username:password@localhost/todo_app
