+++ __init__.py
@@ -1,0 +1,1 @@
+# ディレクトリをPythonパッケージとして認識させるために使用するファイル
